Develop (#3)
* message service added

* env port changed

* Procfile added

* Procfile adde

* env updated

* env updated again

* heroku env updated

* heroku env updated

* heroku env updated 2

* changed vars

* procfile removed

* host changed

* prod config changed

* auth view done

* app refactored

* private messages loaded into pages

* messages appearing on right page only
