+++ public/css/chat.css
@@ -615,7 +615,9 @@ body {
     font-family: "proxima-nova",  "Source Sans Pro", sans-serif;
     float: left;
     border: none;
-    width: calc(100% - 90px);
+
+    width: calc(100% - 0px);
+
     padding: 11px 32px 10px 8px;
     font-size: 0.8em;
     color: #32465a;
