+++ src/app.js
@@ -48,10 +48,12 @@ app.configure(services);
 // Set up event channels (see channels.js)
 app.configure(channels);
 
+app.hooks(appHooks);
+
 // Configure a middleware for 404s and the error handler
 app.use(express.notFound());
 app.use(express.errorHandler({ logger }));
 
-app.hooks(appHooks);
+
 
 module.exports = app;
